requirements.txt
faiss-cpu>=1.9.0
sentence-transformers>=2.2.0
pytest>=7.0.0
pyyaml>=6.0.0
numpy>=1.24.0
